test: added test and refactored them to be the same for every persistence technology
